refactor: refactoring security
