+++ src/main/java/kr/or/cola/backend/config/SecurityConfig.java
@@ -1,6 +1,7 @@
-package kr.or.cola.backend.user.config.auth;
+package kr.or.cola.backend.config;
 
 import kr.or.cola.backend.user.UserService;
+import kr.or.cola.backend.security.OAuth2AuthenticationSuccessHandler;
 import kr.or.cola.backend.user.domain.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
