+++ src/main/java/kr/or/cola/backend/security/dto/SessionUser.java
@@ -1,4 +1,4 @@
-package kr.or.cola.backend.user.config.auth.dto;
+package kr.or.cola.backend.security.dto;
 
 import kr.or.cola.backend.user.domain.Role;
 import kr.or.cola.backend.user.domain.User;
