+++ src/main/java/kr/or/cola/backend/user/domain/User.java
@@ -41,4 +41,5 @@ public class User extends BaseTimeEntity {
     }
 
 
+
 }
