+++ src/main/java/kr/or/cola/backend/userinfo/UserInfoController.java
@@ -1,0 +1,4 @@
+package kr.or.cola.backend.userinfo;
+
+public class UserInfoController {
+}
